SQUASH: Allow not setting DEFAULT_OUTPIN
For targets that do not have an actual output pin (UNIX, or USB serial
port), there is no point in having a DEFAULT_OUTPIN (which is only used
to redirect print #0: "foo" to the right place). I think that this was
previously achieved by not defining HARDWARE_SERIAL_TX, so this commit
might reinstate some previous behaviour (but in the end, they'll get
squashed together again anyway).