SQUASH: Redo the optimizations and blconsole etc. declarations
It turned out the reference trick only worked within a single
compilation unit. Also, the non-optimized versions needed extern in
bitlash.h and an actual definition in a .cpp file (which was exposed by
the UNIX build).

TODO: This commit should be split and then squashed into a few different
commits later on.