Fix UNIX_BUILD by using the ArduinoUnix library
The switch to using Stream and Print classes broke UNIX_BUILD, since no
definitions of those classes were available. Instead of also adding
those to bitlash, this switches to using the ArduinoUnix library and
allows removing all of the unix-specific implementation of the Arduino
API as well.

This requires the ArduinoUnix library to be available (by default in a
directory called `ArduinoUnix` besides the bitlash directory, but this
can be changed by setting ARDUINO_UNIX_PATH).