Restructure the header files
The main goal of this commit is make the build options available to the
public API, since not all functions are available with all options. This
was previously not a problem for most functions (they would be declared
but not defined, which is ok if they're not used either). However, on
Arduino versions < 019, the Stream class is not available, so declaring
initBitlash(Stream&) would break the build, even when it was not used.
Fixing this is the original reason for this commit.

This commit mostly moves declarations around, it should not make any
functional changes. In addition to moving declarations between files, it
also reorganizes the ordering of them and makes comments more
consistent. The major changes in this commit are:

 - The public API used to be in bitlash.h, but is now moved to
   src/bitlash-public.h. bitlash.h still exists for sketches to
   include, but it just includes src/bitlash-public.h.
 - The private API and build config used to be in src/bitlash.h, but is
   now split into src/bitlash-private.h for the private API and
   src/bitlash-config.h for all build config related macros.
   src/bitlash.h no longer exists.
 - src/bitlash-public.h now includes src/bitlash-config.h, to allow the
   public API to be different depending on the build config.
 - A number of functions in the public API are now conditionally
   declared based on the build config. This matches the public API
   declarations to the actual function definitions in the .cpp files.
 - src/bitlash-private.h now includes src/bitlash-public.h. This allows
   removing duplicate declarations for the public API functions and
   types.
 - In the public API, the type of numvar now properly depends on
   TINY_BUILD. Before, the public API would always declare it as long,
   whereas the implementation would declare it as int for TINY_BUILD.
   This will probably have caused TINY_BUILD to be broken.