Various Pinoccio-specific changes
This bulks together all Pinoccio-specific changes previously applied at
https://github.com/Pinoccio/pinoccio-arduino-library. Most of these are
temporary and need better abstractions so they can be removed. It is
expected that upcoming changes to bitlash will remove the need to a lot
of these, so once that happens, we'll restart with a clean bitlash and
properly commit the remaining changes needed separately.